requirements.txt
flask

Flask-SQLAlchemy
Flask-Migrate


